Splitting out cooperative work model
