+++ maproulette/models/__init__.py
@@ -2,3 +2,7 @@ from .project import ProjectModel
 from .challenge import ChallengeModel
 from .task import TaskModel
 from .priority_rule import PriorityRule, PriorityRuleModel
+from .child_operation import ChildOperationModel
+from .parent_operation import ParentOperationModel
+from .cooperative_work import CooperativeWorkModel
+
